Bug fix
Fixed a bug where if you lose and play again you will be greeted with a wave rewards screen at the start of the game.

Changed the ranged unit AI to be smarter and attack enemies in their range when their target dies instead of having to wait for the enemy to leave their range and then come back into it.